+++ tests/integration/src/test/resources/pulsar-messaging.xml
@@ -24,6 +24,8 @@
         <classes>
             <class name="org.apache.pulsar.tests.integration.messaging.PersistentTopicMessagingTest" />
             <class name="org.apache.pulsar.tests.integration.messaging.NonPersistentTopicMessagingTest" />
+            <class name="org.apache.pulsar.tests.integration.proxy.TestProxy" />
+            <class name="org.apache.pulsar.tests.integration.proxy.TestProxyWithWebSocket" />
         </classes>
     </test>
 </suite>
